Move all config files to shared
