+++ server/src/features/player-assignment/padg/utils/assignPadg.ts
@@ -8,7 +8,7 @@ import {
   PadgRandomAssignResults,
 } from "server/typings/padgRandomAssign.typings";
 import { logger } from "server/utils/logger";
-import { getServerConfig } from "server/serverConfig";
+import { getServerConfig } from "shared/config/serverConfig";
 import { calculateHappiness } from "server/features/player-assignment/padg/utils/calculateHappiness";
 import { Input } from "server/typings/result.typings";
