+++ server/src/middleware/cors.ts
@@ -1,6 +1,6 @@
 import { Request, Response, NextFunction } from "express";
 import { logger } from "server/utils/logger";
-import { getServerConfig } from "server/serverConfig";
+import { getServerConfig } from "shared/config/serverConfig";
 
 export const allowCORS = (
   req: Request,
