+++ shared/config/clientConfig.ts
